refactor: navigation, break down presentation components
