+++ app/components/CustomImage.tsx
@@ -4,8 +4,7 @@ export default function CustomImage({ src, alt, ...props }: ImageProps) {
   // For SVGs, do not optimize
   // For other images, use the Next.js Image component normally
   const isSvg = typeof src === 'string' && src.endsWith('.svg')
-
-  // Add a check for the slide assets directory to bypass optimization
+  // But bypass optimization for slide assets
   const isSlideAsset = typeof src === 'string' && src.includes('/slides/')
 
   return (
@@ -14,6 +13,9 @@ export default function CustomImage({ src, alt, ...props }: ImageProps) {
       alt={alt}
       unoptimized={isSvg || isSlideAsset}
       sizes={props.sizes || "(max-width: 768px) 100vw, 1280px"}
+      fetchPriority="high"
+      loading="eager"
+      decoding="async"
       {...props}
     />
   )
