+++ app/components/Slides/Primitives/NavIconNext.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const NavIconNext = () => (
+  <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round">
+    <polyline points="9 18 15 12 9 6" />
+  </svg>
+);
