+++ app/components/Slides/Primitives/NavIconPrev.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+export const NavIconPrev = () => (
+  <svg width="20" height="20" viewBox="0 0 24 24" fill="none" stroke="currentColor" strokeWidth="2.5" strokeLinecap="round" strokeLinejoin="round">
+    <polyline points="15 18 9 12 15 6" />
+  </svg>
+);
